refactor(gui): extract RepoDetail into GitRepoDetail.tsx
